Move code to client package
