+++ src/ir/aminer/potadoshack/Main.java
@@ -1,5 +1,7 @@
 package ir.aminer.potadoshack;
 
+import ir.aminer.potadoshack.client.PotadoShack;
+
 public class Main {
     public static void main(String[] args) {
         javafx.application.Application.launch(PotadoShack.class, args);
