+++ src/ir/aminer/potadoshack/client/controllers/SignIn.java
@@ -1,0 +1,4 @@
+package ir.aminer.potadoshack.client.controllers;
+
+public class SignIn {
+}
