+++ src/ir/aminer/potadoshack/client/controllers/SignUp.java
@@ -1,0 +1,4 @@
+package ir.aminer.potadoshack.client.controllers;
+
+public class SignUp {
+}
