+++ src/ir/aminer/potadoshack/client/layouts/OrderView.fxml
@@ -8,7 +8,7 @@
 
 <AnchorPane xmlns="http://javafx.com/javafx"
             xmlns:fx="http://javafx.com/fxml"
-            fx:controller="ir.aminer.potadoshack.controllers.OrderRegistration"
+            fx:controller="ir.aminer.potadoshack.client.controllers.OrderView"
             prefHeight="400.0" prefWidth="600.0">
 
 </AnchorPane>
