+++ src/ir/aminer/potadoshack/client/layouts/SignIn.fxml
@@ -8,7 +8,7 @@
 
 <AnchorPane xmlns="http://javafx.com/javafx"
             xmlns:fx="http://javafx.com/fxml"
-            fx:controller="ir.aminer.potadoshack.controllers.SignUp"
+            fx:controller="ir.aminer.potadoshack.client.controllers.SignIn"
             prefHeight="400.0" prefWidth="600.0">
 
 </AnchorPane>
