--- src/ir/aminer/potadoshack/controllers/Cart.java
@@ -1,4 +1,0 @@
-package ir.aminer.potadoshack.controllers;
-
-public class Cart {
-}
