--- src/ir/aminer/potadoshack/controllers/MainMenu.java
@@ -1,4 +1,0 @@
-package ir.aminer.potadoshack.controllers;
-
-public class MainMenu {
-}
