--- src/ir/aminer/potadoshack/controllers/SignIn.java
@@ -1,4 +1,0 @@
-package ir.aminer.potadoshack.controllers;
-
-public class SignIn {
-}
