--- src/ir/aminer/potadoshack/controllers/SignUp.java
@@ -1,4 +1,0 @@
-package ir.aminer.potadoshack.controllers;
-
-public class SignUp {
-}
